Users can't delete themselves nor netsle

+++ src/routes/delete.rs
@@ -1,7 +1,7 @@
 use serde::Serialize;
 use rocket::request::Form;
 use rocket::response::content;
-use rocket::http::Status;
+use rocket::response::status::BadRequest;
 use crate::database;
 use crate::guards::{UserDeleteForm, ApiKey, Admin};
 use crate::MainDbConn;
@@ -13,11 +13,15 @@ struct ErrorResponse {
 
 #[delete("/admin/delete_user", data = "<delete_form>")]
 pub fn delete_user(
-    _key: ApiKey,
+    key: ApiKey,
     _admin: Admin,
     conn: MainDbConn,
     delete_form: Form<UserDeleteForm>
-) -> Result<content::Json<String>, Status> {
+) -> Result<content::Json<String>, BadRequest<&'static str>> {
+    if key.0.claims.sub == delete_form.username || delete_form.username == "netsle" {
+        return Err(BadRequest(Some("{\"message\": \"Can't delete yourself nor netsle you silly goose!\"}")));
+    }
+
     let result = database::users::delete_user(&*conn, &delete_form.username);
 
     match result {
@@ -25,7 +29,7 @@ pub fn delete_user(
             Ok(content::Json("{}".to_string()))
         },
         Err(_) => {
-            Err(Status::new(500, "Something bad happened with our database :)"))
+            Err(BadRequest(Some("{\"message\": \"Something bad happened with our database :)\"}")))
         }
     }
 }
